% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSimilar.R
\name{compareWithFingerprint}
\alias{compareWithFingerprint}
\title{Compare a test fingerprint with a known fingerprint}
\usage{
compareWithFingerprint(test, fingerprint, threshold = 0.001, exact = FALSE)
}
\arguments{
\item{test}{fingerprint}

\item{fingerprint}{fingerprint}

\item{threshold}{single numeric similarity parameter,
for the \code{original} algorithm it defaults to 1e-3.
For the \code{dcf} method it is the number bits allowed to
be different in the fingerprint (i.e. Hamming distance).}

\item{exact}{single logical should fingerprints match exactly (default
\code{FALSE})}
}
\value{
single logical
}
\description{
Compare a test fingerprint with a known fingerprint
}
\examples{
compareWithFingerprint(test = 1:3, fingerprint = 1:3)
compareWithFingerprint(test = 1:3, fingerprint = 1:4)
compareWithFingerprint(test = 1:3, fingerprint = 1:3 + 1e-3)
compareWithFingerprint(test = 1:3, fingerprint = c(1, 2, 3.1))
compareWithFingerprint(
  test = 1:3,
  fingerprint = c(1, 2, 3.1),
  exact = TRUE
)
}

