% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFingerprint.R
\name{getFingerprint}
\alias{getFingerprint}
\title{Get Image Fingerprint}
\usage{
getFingerprint(file, algorithm = c("dcf", "original"))
}
\arguments{
\item{file}{single character naming PNG, JPG or BMP file
from which to get fingerprint. It can also be a gzip compressed
file with extension \code{.gz}.}

\item{algorithm}{hashing algorithm. Possible values:
\code{original}, \code{dcf}. See details below.}
}
\description{
TODO: details of the algorithms.
}
\examples{
getFingerprint(
  system.file(package = "visualTest", "compare", "stest-00.jpg.gz")
)
}

