% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFingerprint.R
\name{isCross}
\alias{isCross}
\title{Find Zero Crossing Points}
\usage{
isCross(x, len = 3)
}
\arguments{
\item{x}{numeric vector}

\item{len}{natural number window length for sign comparison}
}
\value{
logical vector of length zero
}
\description{
Find positions in a numeric vector where sign has changed on average for
more than n elements.
}
\examples{
x1 <- -7:8
visualTest:::isCross(x = x1)
x2 <- rep(-1:1, times = 6)
visualTest:::isCross(x = x2)
}
\keyword{internal}

