% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSimilar.R
\name{isSimilar}
\alias{isSimilar}
\title{Is this Graphic a Match to this Fingerprint?}
\usage{
isSimilar(file, fingerprint, threshold = 0.001, exact = FALSE, ...)
}
\arguments{
\item{file}{single character naming PNG or JPG file from which to get
fingerprint}

\item{fingerprint}{fingerprint, or filename. Character scalars
without a dot in them and numeric vectors are taken as fingerprints.}

\item{threshold}{single numeric similarity parameter (default 1e-3)}

\item{exact}{single logical should fingerprints match exactly (default
\code{FALSE}). Note that for the \code{original} algorithm, this
is not equivalent to using a zero threshold.}

\item{...}{additional arguments}
}
\description{
A function to check similarity between plots. By default, fingerprint
comparison is very fuzzy, and the fuzziness can be controlled by
argument threshold (see \code{\link{compareWithFingerprint}}).
Fingerprint exact matching can be selected with argument exact. Note
that similar looking images may have the same fingerprint (see
\code{\link{getFingerprint}}).
}
\examples{
rdata <- matrix(rnorm(200), ncol = 2)

## To create the images in a temporary directory
tmp <- function(path) file.path(tempdir(), path)

png(tmp("test1.png"))
plot(rdata)
dev.off()

fing1 <- getFingerprint(file = tmp("test1.png"))
rdata[2, 2] <- 0.1
png(tmp("test2.png"))
plot(rdata)
dev.off()

isSimilar(file = tmp("test2.png"), fingerprint = fing1)
isSimilar(file = tmp("test2.png"), fingerprint = fing1, threshold = 0.05)

png(tmp("test3.png"))
plot(rdata, col = 3)
dev.off()

isSimilar(file = tmp("test3.png"), fingerprint = fing1)
isSimilar(file = tmp("test3.png"), fingerprint = fing1, threshold = 0.05)

rdata2 <- matrix(rnorm(200), ncol = 2)
png(tmp("test4.png"))
plot(rdata2)
dev.off()

isSimilar(file = tmp("test4.png"), fingerprint = fing1)
isSimilar(file = tmp("test4.png"), fingerprint = fing1, threshold = 5)

png(tmp("test5.png"))
hist(rdata2)
dev.off()

isSimilar(file = tmp("test5.png"), fingerprint = fing1)
isSimilar(file = tmp("test5.png"), fingerprint = fing1, threshold = 5e6)
}

