% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFingerprint.R
\name{rgb2Value}
\alias{rgb2Value}
\title{Convert RGB array to value}
\usage{
rgb2Value(array)
}
\arguments{
\item{array}{array with three dimensions length N, M and 3}
}
\value{
matrix with nrow N and ncol M
}
\description{
Convert RGB array to value
}
\keyword{internal}

