% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFingerprint.R
\name{showFingerprint}
\alias{showFingerprint}
\title{Plot some fingerprints against each other}
\usage{
showFingerprint(...)
}
\arguments{
\item{...}{files to test}
}
\value{
list of fingerprint(s) invisibly.
}
\description{
Plot some fingerprints against each other
}
\examples{
sf <- system.file(package = "visualTest")
eg <- "VR-616_plot-lm00.jpg.gz"
showFingerprint(
  file.path(sf, "compare", "windows", eg),
  file.path(sf, "compare", "unix", eg)
)
}

