% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSimilar.R
\docType{package}
\name{visualTest-package}
\alias{visualTest}
\alias{visualTest-package}
\title{Extract and match characteristic signal in graphs}
\description{
Use \pkg{visualTest} to perform fuzzy image matching
}
\author{
Chris Campbell \email{ccampbell@mango-solutions.com}
}

