% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFingerprint.R
\name{getFingerprint}
\alias{getFingerprint}
\title{Get Image Fingerprint}
\usage{
getFingerprint(file, ...)
}
\arguments{
\item{file}{single character naming PNG, JPG or BMP file
from which to get fingerprint}

\item{...}{additional arguments}
}
\description{
Get Image Fingerprint
}
\examples{
getFingerprint(
  system.file(package = "visualTest", "compare", "stest-00.jpg")
)
}

