% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFingerprint.R
\name{showFingerprint}
\alias{showFingerprint}
\title{Plot some fingerprints against each other
This currently only works for fingerprints from the
\sQuote{original} algorithm.}
\usage{
showFingerprint(..., algorithm = "original")
}
\arguments{
\item{...}{files to test}

\item{algorithm}{The algorithm to use, see
\code{\link{getFingerprint}}. Currently only the \sQuote{original}
algorithm works.}
}
\value{
list of fingerprint(s) invisibly.
}
\description{
Plot some fingerprints against each other
This currently only works for fingerprints from the
\sQuote{original} algorithm.
}
\examples{
sf <- system.file(package = "visualTest")
eg <- "VR-616_plot-lm00.jpg.gz"
showFingerprint(
  file.path(sf, "compare", "windows", eg),
  file.path(sf, "compare", "unix", eg)
)
}

