\docType{package}
\name{visualTest-package}
\alias{visualTest}
\alias{visualTest-package}
\title{Extract and match characteristic signal in graphs}
\description{
Use \pkg{visualTest} to perform fuzzy image matching
}
\details{
\tabular{ll}{ Package: \tab visualTest \cr Type: \tab
Package \cr Version: \tab 0.1 \cr Date: \tab April 2014 \cr
Lazyload: \tab yes \cr }
}
\author{
ccampbell \email{ccampbell@mango-solutions.com}
}

